# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "NHTSA API",
        "contact": {
            "name": "Marco Araujo",
            "url": "http://www.marcojunior.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
        "/vehicles/{year}/{manufacturer}/{model}": {
            "get": {
                "tags": [
                    "vehicles"
                ],
                "summary": "Display a listing of the vehicles by Model Year, Manufacturer and Model.",
                "description": "Returns list of vehicles",
                "operationId": "vehicles.findAllByAttributes",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "year",
                        "in": "path",
                        "description": "The model year of vehicle",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "manufacturer",
                        "in": "path",
                        "description": "The manufacturer of vehicle",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "model",
                        "in": "path",
                        "description": "The model of vehicle",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "withRating",
                        "in": "query",
                        "description": "Determines whether show crash rating",
                        "required": false,
                        "type": "boolean",
                        "default": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of vehicles",
                        "schema": {
                            "properties": {
                                "Counts": {
                                    "description": "Total results",
                                    "type": "integer",
                                    "example": 1
                                },
                                "Results": {
                                    "description": "Display a listing of the vehicles by Model Year, Manufacturer and Model.",
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "Description": {
                                                "description": "Description of vehicle",
                                                "type": "string",
                                                "example": "2015 Audi A3 C FWD"
                                            },
                                            "VehicleId": {
                                                "description": "Id of vehicle",
                                                "type": "integer",
                                                "example": 9406
                                            },
                                            "CrashRating": {
                                                "description": "Crash Rating of vehicle (is shown when withRating is true)",
                                                "type": "string",
                                                "example": "Not Rated"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/vehicles": {
            "post": {
                "tags": [
                    "vehicle"
                ],
                "summary": "Create vehicle",
                "description": "Create a vehicle record",
                "operationId": "create",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Created vehicle object",
                        "required": true,
                        "schema": {
                            "properties": {
                                "modelYear": {
                                    "type": "integer"
                                },
                                "manufacturer": {
                                    "type": "string"
                                },
                                "model": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "definitions": {
        "Error": {
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "description": "Class Controller",
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "description": "Class Controller",
                    "type": "string"
                }
            }
        }
    }
}